#2327 fix btrfs folder teardown in tests

+++ test/unittests/libethereum/ClientTest.cpp
@@ -104,7 +104,8 @@ struct FixtureCommon {
 #endif
     }
 
-    void cleanupBtrfsArtifacts( const std::string& _mountPath, const std::string& _imagePath ) {
+    void cleanupBtrfsArtifacts(
+        const std::string& _mountPath, const std::string& _imagePath, bool _removeMountPath ) {
         gainRoot();
 #if ( !defined __APPLE__ )
         while ( system( ( "mountpoint -q " + _mountPath ).c_str() ) == 0 ) {
@@ -112,14 +113,17 @@ struct FixtureCommon {
             assert( rv == 0 );
         }
 #endif
-        int rv = system( ( "rm -rf " + _mountPath ).c_str() );
-        assert( rv == 0 );
+        int rv;
+        if ( _removeMountPath ) {
+            rv = system( ( "rm -rf " + _mountPath ).c_str() );
+            assert( rv == 0 );
+        }
         rv = system( ( "rm -f " + _imagePath ).c_str() );
         assert( rv == 0 );
     }
 
     void cleanupBtrfsArtifacts() {
-        cleanupBtrfsArtifacts( BTRFS_DIR_PATH, BTRFS_FILE_PATH );
+        cleanupBtrfsArtifacts( BTRFS_DIR_PATH, BTRFS_FILE_PATH, true );
     }
 };
 
@@ -266,13 +270,12 @@ class TestClientSnapshotsFixture : public TestOutputHelperFixture, public Fixtur
 public:
     TestClientSnapshotsFixture( const std::string& _config ) try {
         check_sudo();
-        cleanupBtrfsArtifacts( m_tmpDir.path(), BTRFS_FILE_PATH );
+        cleanupBtrfsArtifacts( m_tmpDir.path(), BTRFS_FILE_PATH, false );
 
         dropRoot();
 
         int rv = system( ( "dd if=/dev/zero of=" + BTRFS_FILE_PATH + " bs=1M count=200" ).c_str() );
         rv = system( ( "mkfs.btrfs " + BTRFS_FILE_PATH ).c_str() );
-        rv = system( ( "mkdir " + m_tmpDir.path() ).c_str() );
 
         gainRoot();
         rv =
@@ -360,7 +363,7 @@ public:
         const char* NC = getenv( "NC" );
         if ( NC )
             return;
-        cleanupBtrfsArtifacts( m_tmpDir.path(), BTRFS_FILE_PATH );
+        cleanupBtrfsArtifacts( m_tmpDir.path(), BTRFS_FILE_PATH, false );
     }
 
 private:
